Add --friendly-names command-line boolean flag
If the flag is set, the module name is used as the base for the caching directory and for the file name.  It makes querying what is actually in the cache possible just by folder listing.  As long as all modules come from npmregistry.org, there should be no name clash.