Improve caching of tgz files
The main idea is to cache the .tgz files and never delete the cache. The reason is that the .tgz files from registry cannot be changed.
However during download of a specific .tgz the file may not be fully downloaded. So we may get incorrect file in the cache. In case when the .tgz file exist in the proxy cache, we verify it by checking:
 - try reading the file with information about the package. This is the full file from `http://registry.npmjs.org/<package_name>`. It contains information about the shasum of each version's tgz. Get the shasum of our .tgz and compare it agains the one in the file. In case they are the same, use the file. In any other case download it again.
 - in case the file that contains information about shasum of package does not exist, try getting the shasum by spawning npm. Give it 10 seconds and if the shasum is returned, compare it with our .tgz shasum.

Add maximum waiting for lock of the files. During writing the files to the proxy cache, a lock is created. If during this time there are any subsequent requests for this package, they are blocked until the lock is released. In some case npm hangs and the lock remains forever. Raise error after 50 seconds. This will kill the process, but the daemon will restart it. Nom will retry the operation (as the previous one will return error) and most probably it will succeed.