Return existing tgz in case npm view hangs
In case npm view command hangs, we exit it after 10 seconds. Instead of downloading the .tgz again, return the already existing one.
Npm client will verify it on the other side. In case we try to download the .tgz again, there's a very good chance to hang as well (npm view calls the registry as well).